Abstract input related components in Stake and Unstake page to StakingInput
